Req Changes as minor refactor; "native structs" together in 1 folder.
