+++ AssetManagment/Strategies/Fbx/DllDefinitions/FBXSCeneContainerDll.cs
@@ -7,7 +7,6 @@ namespace AssetManagement.Strategies.Fbx.DllDefinitions
     {
         const String dllFileName = "FBXWrapperNative.dll";
 
-
         [DllImport(dllFileName, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool GetPackedVertices(IntPtr ptrInstances, int meshIndex, out IntPtr vertices, out int itemCount);
