+++ AssetManagment/Strategies/Fbx/SceneMarshallerToManaged.cs
@@ -1,5 +1,6 @@
 ﻿using AssetManagement.GenericFormats;
-using AssetManagement.GenericFormats.Native;
+using AssetManagement.GenericFormats.Managed;
+using AssetManagement.GenericFormats.Unmanaged;
 using AssetManagement.Strategies.Fbx.DllDefinitions;
 using System;
 using System.Collections.Generic;
